requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
cffi==1.12.3
chardet==3.0.4
discord.py==1.2.5
dnspython==1.16.0
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
numpy==1.16.3
pandas==0.24.2
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
typing-extensions==3.7.2
websockets==6.0
yarl==1.3.0
youtube-dl==2019.12.25
